Move several TLS key/certificate loading routines into a util module
